wayland: xdg_surface_configure(): call wl_surface_commit()
It seems kwin expects a wl_surface_commit() for each
xdg_surface_ack_configure().

We don't want to commit before we've rendered a resized surface. So,
if we *did* change the window size, let the normal rendering path do
the surface commit.

Only when we did *not* change the window size do we need to explicitly
commit the surface in xdg_surface_configure().